+++ src/main/java/com/AhmedNaeim/assignment3/Server/hotelService.java
@@ -38,7 +38,7 @@ public class hotelService {
         return booking;
     }
 
-    public List<bookings> getRoomBookings(int id) {
+    public List<bookings> getRoomBookings(long id) {
         final List<bookings> bookingsList = new ArrayList<>();
 
         for (int i = 0; i < booking.size(); i++) {
